+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Output a warning when a stack overflow has been detected in `px4_tasks` command.
 - Configure GDB to display absolute filenames and one line of disassembly.
 - Show function and file line in callgraph to easier distinguish nodes.
+- Add malloc fuzzer to find code without proper OOM handling.
 
 ## 1.1.1